PHPunit was failing, due to db migrations and incorrectly casting integers

+++ database/migrations/2018_09_25_152421_delete_H_NMR_C_NMR_boolean_columns_from_compounds_table.php
@@ -13,10 +13,12 @@ class DeleteHNMRCNMRBooleanColumnsFromCompoundsTable extends Migration
      */
     public function up()
     {
-        Schema::table('compounds', function (Blueprint $table) {
-            $table->dropColumn('proton_nmr');
-            $table->dropColumn('carbon_nmr');
-        });
+        if(Schema::hasColumn('proton_nmr', 'carbon_nmr')) {
+            Schema::table('compounds', function (Blueprint $table) {
+                $table->dropColumn('proton_nmr');
+                $table->dropColumn('carbon_nmr');
+            });    
+        }
     }
 
     /**
